refactor: address PR review feedback on Umami widget
- Split component.tsx into focused files: umami-content.tsx,
  umami-events-content.tsx, umami-top-list.tsx, umami-utils.ts
- Replace local formatDuration with @homarr/common's implementation
- Fix timezone: use local time instead of UTC for chart axis labels
- Fix null referrer: umamiMetricItemSchema.x now accepts null (direct
  traffic) and normalises it to empty string via Zod transform
- Extract UMAMI_CUSTOM_EVENT_TYPE constant (was magic number 2)
- Simplify formatTimeFrameLabel switch to template literal lookup
- Add comment on computeTimeRange server-timezone limitation
- Add missing tests for getTopPagesAsync, getTopReferrersAsync,
  and getMultiEventTimeSeriesAsync including null referrer case
- Fix Prettier formatting violations
